requirements.txt
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
gradio>=3.50.0
pillow>=9.0.0
torchvision>=0.15.0
numpy>=1.24.0
tqdm>=4.65.0
accelerate>=0.20.0